+++ tests/phpunit/PDO/PDOQueryInterfaceTest.php
@@ -98,7 +98,6 @@ class PDOQueryInterfaceTest extends \PHPUnit_Framework_TestCase {
 			'UpdateSqlBuilder' => $this->getMock( 'Wikibase\Database\QueryInterface\UpdateSqlBuilder' ),
 			'DeleteSqlBuilder' => $this->getMock( 'Wikibase\Database\QueryInterface\DeleteSqlBuilder' ),
 			'SelectSqlBuilder' => $this->getMock( 'Wikibase\Database\QueryInterface\SelectSqlBuilder' ),
-			'InsertedIdSqlBuilder' => $this->getMock( 'Wikibase\Database\QueryInterface\InsertedIdSqlBuilder' ),
 		);
 
 		$sqlBuilders[$collaboratorName] = $collaboratorInstance;
@@ -107,15 +106,13 @@ class PDOQueryInterfaceTest extends \PHPUnit_Framework_TestCase {
 		$updateBuilder = $sqlBuilders['UpdateSqlBuilder'];
 		$deleteBuilder = $sqlBuilders['DeleteSqlBuilder'];
 		$selectBuilder = $sqlBuilders['SelectSqlBuilder'];
-		$idBuilder = $sqlBuilders['InsertedIdSqlBuilder'];
 
 		return new PDOQueryInterface(
 			$pdo,
 			$insertBuilder,
 			$updateBuilder,
 			$deleteBuilder,
-			$selectBuilder,
-			$idBuilder
+			$selectBuilder
 		);
 	}
 
@@ -182,31 +179,5 @@ class PDOQueryInterfaceTest extends \PHPUnit_Framework_TestCase {
 
 		// TODO: assert contents
 	}
-
-	protected function newQueryInterfaceForGetInsertId( $getIdCallReturnValue ) {
-		$insertedIdBuilder = $this->getMock( 'Wikibase\Database\QueryInterface\InsertedIdSqlBuilder' );
-
-		$insertedIdBuilder->expects( $this->once() )
-			->method( 'getSqlToGetTheInsertedId' )
-			->will( $this->returnValue( $this->stubSql ) );
-
-		return $this->newQueryInterface( $getIdCallReturnValue, 'InsertedIdSqlBuilder', $insertedIdBuilder );
-	}
-
-	public function testGetInsertIdCallsCollaboratorsAndReturnsInt() {
-		$db = $this->newQueryInterfaceForGetInsertId( '1337' );
-		$id = $db->getInsertId();
-
-		$this->assertInternalType( 'int', $id );
-		$this->assertEquals( 1337, $id );
-	}
-
-	public function testOnReceiveOfFalse_getInsertIdThrows() {
-		$db = $this->newQueryInterfaceForGetInsertId( false );
-
-		$this->setExpectedException( 'Wikibase\Database\QueryInterface\QueryInterfaceException' );
-		$db->getInsertId();
-	}
-
 
 }